refactor: extract validation functions to shared utilities module
